+++ src/Custom/FineTuning/Internal/Pagination/FineTuningJobEventsPageEnumerator.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.ClientModel;
 using System.ClientModel.Primitives;
-using System.Collections.Generic;
 using System.Text.Json;
 using System.Threading.Tasks;
 
@@ -95,7 +94,7 @@ internal partial class FineTuningJobEventsPageEnumerator : PageResultEnumerator
         request.Method = "GET";
         var uri = new ClientUriBuilder();
         uri.Reset(_endpoint);
-        uri.AppendPath("/fine_tuning/jobs/", false);
+        uri.AppendPath("/v1/fine_tuning/jobs/", false);
         uri.AppendPath(fineTuningJobId, true);
         uri.AppendPath("/events", false);
         if (after != null)